db_safe_layer/requirements.txt
sqlglot>=25.6.0
SQLAlchemy>=2.0.30
pydantic>=2.8.2
python-dotenv>=1.0.1
openai>=1.40.0
pandas>=2.2.2
tabulate>=0.9.0
flask>=2.3.0
psycopg2-binary>=2.9.0
